chore(rust): factor out some utils into polars-time/src/utils
